@@ -11,21 +11,21 @@
         "php": "^7.4",
         "ext-json": "*",
         "bugsnag/bugsnag-laravel": "^2.0",
-        "facade/ignition": "^2.0",
+        "facade/ignition": "^2.3.6",
         "fideloper/proxy": "^4.0",
-        "guzzlehttp/guzzle": "^7.0",
-        "laravel/framework": "^7.0",
+        "guzzlehttp/guzzle": "^7.0.1",
+        "laravel/framework": "^8.0",
         "laravel/tinker": "^2.0",
-        "laravel/ui": "^2.0",
+        "laravel/ui": "^3.0",
         "php-ffmpeg/php-ffmpeg": "^0.16.0"
     },
     "require-dev": {
         "barryvdh/laravel-debugbar": "^3.4",
         "fzaninotto/faker": "^1.4",
         "laravel/dusk": "^6.4",
-        "laravel/telescope": "^3.0",
+        "laravel/telescope": "^4.0",
         "mockery/mockery": "^1.0",
-        "nunomaduro/collision": "^4.1",
+        "nunomaduro/collision": "^5.0",
         "phpunit/phpunit": "^9.2"
     },
     "config": {
@@ -40,12 +40,10 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "app/"
-        },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
