@@ -11,80 +11,119 @@
 |
 */
 
+use App\Http\Controllers\AdminController;
+use App\Http\Controllers\CardController;
+use App\Http\Controllers\CourseController;
+use App\Http\Controllers\EnrollmentController;
+use App\Http\Controllers\FileController;
+use App\Http\Controllers\FolderController;
+use App\Http\Controllers\InvitationController;
+use App\Http\Controllers\LocalizationController;
+use App\Http\Controllers\UserController;
+
 // Authentication routes
 Auth::routes();
 
 // SwitchAAI authentication route
-Route::get('aai', 'CourseController@index')->name('aai')->middleware('check_aai');
+Route::get('aai', [CourseController::class, 'index'])
+    ->name('aai')
+    ->middleware('check_aai');
 
 // Localization route
-Route::get('lang/{locale}', 'LocalizationController@index');
+Route::get('lang/{locale}', [LocalizationController::class, 'index']);
 
 // Register with an invitation routes
 Route::middleware(['has_invitation'])->group(function () {
-    Route::get('invitations/register', 'InvitationController@register')->name('invitations.register');
-    Route::post('invitations/user/create', 'InvitationController@createInvitationUser')->name('invitations.user.create');
+    Route::get('invitations/register', [InvitationController::class, 'register'])
+        ->name('invitations.register');
+    Route::post('invitations/user/create', [InvitationController::class, 'createInvitationUser'])
+        ->name('invitations.user.create');
 });
 
 // Standard routes for authenticated users
 Route::middleware(['auth', 'app'])->group(function () {
-    Route::get('/', 'CourseController@index')->name('home');
+    // Homepage
+    Route::get('/', [CourseController::class, 'index'])
+        ->name('home');
 
     // Invitations
-    Route::resource('invitations', 'InvitationController');
-    Route::get('invitations/{invitation}/mail', 'InvitationController@mail')->name('send.invite');
+    Route::resource('invitations', InvitationController::class);
+    Route::get('invitations/{invitation}/mail', [InvitationController::class, 'mail'])
+        ->name('send.invite');
 
     // Users
-    Route::get('users/{user}/profile', 'UserController@profile')->name('users.profile');
-    Route::put('users/{user}/update', 'UserController@update')->name('users.profile.update');
+    Route::get('users/{user}/profile', [UserController::class, 'profile'])
+        ->name('users.profile');
+    Route::put('users/{user}/update', [UserController::class, 'update'])
+        ->name('users.profile.update');
 
     // Cards
-    Route::resource('cards', 'CardController');
-    Route::put('cards/{card}/unlink/file', 'CardController@unlinkFile')->name('cards.unlink.file');
-    Route::put('cards/{card}/editor', 'CardController@editor')->name('cards.editor');
+    Route::resource('cards', CardController::class);
+    Route::put('cards/{card}/unlink/file', [CardController::class, 'unlinkFile'])
+        ->name('cards.unlink.file');
+    Route::put('cards/{card}/editor', [CardController::class, 'editor'])
+        ->name('cards.editor');
 
     // Folders
-    Route::resource('folders', 'FolderController');
+    Route::resource('folders', FolderController::class);
 
     // Files
-    Route::resource('files', 'FileController')->only(['index', 'destroy']);
-    Route::post('files/upload', 'FileController@upload');
+    Route::resource('files', FileController::class)->only([
+        'index',
+        'destroy'
+    ]);
+    Route::post('files/upload', [FileController::class, 'upload']);
 
     // Enrollments
-    Route::get('enrollments', 'EnrollmentController@index');
-    Route::post('enrollments', 'EnrollmentController@store');
-    Route::get('enrollments/find', 'EnrollmentController@find');
-    Route::put('enrollments/cards', 'EnrollmentController@cards');
-    Route::delete('enrollments/{enrollment}', 'EnrollmentController@destroy');
+    Route::get('enrollments', [EnrollmentController::class, 'index']);
+    Route::post('enrollments', [EnrollmentController::class, 'store']);
+    Route::get('enrollments/find', [EnrollmentController::class, 'find']);
+    Route::put('enrollments/cards', [EnrollmentController::class, 'cards']);
+    Route::delete('enrollments/{enrollment}', [EnrollmentController::class, 'destroy']);
 
     // Courses
-    Route::get('courses/{course}', 'CourseController@show')->name('courses.show');
-    Route::get('courses/{course}/configure', 'CourseController@configure')->name('courses.configure');
-    Route::get('courses/{course}/configure/files', 'FileController@index')->name('courses.configure.files');
-    Route::put('/courses/{course}/archive', 'CourseController@archive')->name('courses.archive');
-    Route::delete('/courses/{course}/disable', 'CourseController@disable')->name('courses.disable');
+    Route::get('courses/{course}', [CourseController::class, 'show'])
+        ->name('courses.show');
+    Route::get('courses/{course}/configure', [CourseController::class, 'configure'])
+        ->name('courses.configure');
+    Route::get('courses/{course}/configure/files', [FileController::class, 'index'])
+        ->name('courses.configure.files');
+    Route::put('/courses/{course}/archive', [CourseController::class, 'archive'])
+        ->name('courses.archive');
+    Route::delete('/courses/{course}/disable', [CourseController::class, 'disable'])
+        ->name('courses.disable');
 });
 
 // Administration routes
 Route::group(['prefix' => 'admin',  'as' => 'admin.', 'middleware' => ['auth', 'app', 'is_admin']], function() {
-    Route::get('/', 'AdminController@index')->name('index');
+    // Admin homepage
+    Route::get('/', [AdminController::class, 'index'])
+        ->name('index');
 
     // Invitations
-    Route::get('/invitations', 'InvitationController@manage')->name('invitations.manage');
-    Route::get('/invitations/create', 'InvitationController@create')->name('invitations.create');
+    Route::get('/invitations', [InvitationController::class, 'manage'])
+        ->name('invitations.manage');
+    Route::get('/invitations/create', [InvitationController::class, 'create'])
+        ->name('invitations.create');
 
     // Users
-    Route::resource('users', 'UserController');
-    Route::get('/users', 'UserController@manage')->name('users.manage');
-    Route::get('/users/{user}/extend', 'UserController@extend')->name('users.extend');
+    Route::resource('users', UserController::class);
+    Route::get('/users', [UserController::class, 'manage'])
+        ->name('users.manage');
+    Route::get('/users/{user}/extend', [UserController::class, 'extend'])
+        ->name('users.extend');
 
     // Courses
-    Route::resource('courses', 'CourseController');
-    Route::get('/courses', 'CourseController@manage')->name('courses.manage');
-    Route::get('/courses/{course}/enable', 'CourseController@enable')->name('courses.enable');
-    Route::get('/courses/{course}/mailConfirmDelete', 'CourseController@mailConfirmDelete')->name('courses.send.confirm.delete');
+    Route::resource('courses', CourseController::class);
+    Route::get('/courses', [CourseController::class, 'manage'])
+        ->name('courses.manage');
+    Route::get('/courses/{course}/enable', [CourseController::class, 'enable'])
+        ->name('courses.enable');
+    Route::get('/courses/{course}/mailConfirmDelete', [CourseController::class, 'mailConfirmDelete'])
+        ->name('courses.send.confirm.delete');
 
     // Files
-    Route::resource('files', 'FileController');
-    Route::get('/files', 'FileController@manage')->name('files.manage');
+    Route::resource('files', FileController::class);
+    Route::get('/files', [FileController::class, 'manage'])
+        ->name('files.manage');
 });
