@@ -1,41 +1,64 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
+
 use App\User;
-use Faker\Generator as Faker;
+use Carbon\Carbon;
+use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Support\Str;
 
-/*
-|--------------------------------------------------------------------------
-| Model Factories
-|--------------------------------------------------------------------------
-|
-| This directory should contain each of the model factory definitions for
-| your application. Factories provide a convenient way to generate new
-| model instances for testing / seeding your application's database.
-|
-*/
+class UserFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = User::class;
 
-$factory->define(User::class, function (Faker $faker) {
-    return [
-        'name' => $faker->name,
-        'email' => $faker->unique()->safeEmail,
-        'email_verified_at' => now(),
-        'password' => '$2y$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi',
-        'remember_token' => Str::random(10),
-    ];
-});
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'name' => $this->faker->name,
+            'email' => $this->faker->unique()->safeEmail,
+            'email_verified_at' => now(),
+            'password' => '$2y$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // password
+            'remember_token' => Str::random(10),
+        ];
+    }
 
-$factory->state(User::class, 'admin', function() {
-    return [
-        'admin' => true,
-    ];
-});
+    /**
+     * Indicate that the user is an admin.
+     *
+     * @return Factory
+     */
+    public function admin()
+    {
+        return $this->state(function (array $attributes) {
+            return [
+                'admin' => true,
+            ];
+        });
+    }
 
-$factory->state(User::class, 'invalid', function() {
-    $now = Carbon\Carbon::now();
+    /**
+     * Indicate that the user has an invalid account.
+     *
+     * @return Factory
+     */
+    public function invalid()
+    {
+        $now = Carbon::now();
 
-    return [
-        'validity' => $now->subDays(1),
-    ];
-});
+        return $this->state(function (array $attributes) use ($now) {
+            return [
+                'validity' => $now->subDays(1),
+            ];
+        });
+    }
+}
